backend-fastapi/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
python-dotenv
aiosqlite
faiss-cpu
sentence-transformers
numpy
langchain
langchain-groq